requirements.txt
flask
flask-cors
anthropic
python-dotenv
google-genai
Pillow
